Transfered the table for fetching data of assistants in the assistants page
